Added side menu plus some cleaning up in files stracture
