+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ArticleView" />
+        <activity android:name=".activity.ArticleView" />
         <activity android:name=".FilterTeams"></activity>
     </application>
