+++ app/src/main/res/layout/activity_article_view.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.yavengy.hidemyteam.ArticleView">
+    tools:context="com.example.yavengy.hidemyteam.activity.ArticleView">
 
     <WebView
         android:layout_width="match_parent"
