--- app/src/main/res/layout/activity_filter_teams.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_filter_teams"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.yavengy.hidemyteam.FilterTeams">
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="64dp"
-        android:clipToPadding="false"
-        android:scrollbars="vertical" />
-
-    <include
-        android:id="@+id/toolbar"
-        layout="@layout/toolbar" />
-
-</RelativeLayout>
