+++ app/src/main/res/layout/activity_main.xml
@@ -6,36 +6,22 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-<RelativeLayout
+<LinearLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/viewBg"
-    tools:context="com.example.yavengy.hidemyteam.MainActivity">
+    android:orientation="vertical"
+    tools:context="com.example.yavengy.hidemyteam.activity.MainActivity">
 
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_view"
+    <LinearLayout
+        android:id="@+id/container_toolbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="64dp"
-        android:clipToPadding="false"
-        android:scrollbars="vertical" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:src="@drawable/arrow_down_circle"
-        android:layout_marginLeft="@dimen/fab_margin"
-        android:layout_marginRight="@dimen/fab_margin"
-        android:layout_marginBottom="68dp"
-        android:layout_marginEnd="32dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true" />
-
-    <include
-        android:id="@+id/toolbar"
-        layout="@layout/toolbar" />
+        android:orientation="vertical">
+        <include
+            android:id="@+id/toolbar"
+            layout="@layout/toolbar" />
+    </LinearLayout>
 
     <FrameLayout
         android:id="@+id/container_body"
@@ -43,7 +29,7 @@
         android:layout_height="0dp"
         android:layout_weight="1" />
 
-</RelativeLayout>
+</LinearLayout>
 
     <fragment
         android:id="@+id/fragment_navigation_drawer"
