+++ app/src/main/res/layout/fragment_home.xml
@@ -1,20 +1,17 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.yavengy.hidemyteam.activity.HomeFragment">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/viewBg"
-        tools:context="com.example.yavengy.hidemyteam.MainActivity">
+    android:orientation="vertical"
+    tools:context="com.example.yavengy.hidemyteam.activity.HomeFragment"
+    android:weightSum="1">
 
         <android.support.v7.widget.RecyclerView
             android:id="@+id/recycler_view"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="64dp"
+            android:layout_height="match_parent"
+            android:layout_marginTop="5dp"
+            android:layout_marginBottom="5dp"
             android:clipToPadding="false"
             android:scrollbars="vertical" />
 
@@ -31,16 +28,4 @@
             android:layout_alignParentBottom="true"
             android:layout_alignParentEnd="true" />
 
-        <include
-            android:id="@+id/toolbar"
-            layout="@layout/toolbar" />
-
-        <FrameLayout
-            android:id="@+id/container_body"
-            android:layout_width="fill_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1" />
-
-    </RelativeLayout>
-
-</FrameLayout>
+</RelativeLayout>
