+++ app/src/main/res/menu/menu_filter.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".FilterTeams">
+    tools:context=".FilterFragment">
 
     <item
         android:id="@+id/action_back"
