+++ app/src/main/res/values/strings.xml
@@ -6,18 +6,18 @@
     <string name="drawer_close">Close</string>
 
     <string name="nav_item_home">Home</string>
-    <string name="nav_item_friends">Friends</string>
-    <string name="nav_item_notifications">Messages</string>
+    <string name="nav_item_filters">Filter</string>
+    <string name="nav_item_skin">Change Skin</string>
 
     <!-- navigation drawer item labels  -->
     <string-array name="nav_drawer_labels">
         <item>@string/nav_item_home</item>
-        <item>@string/nav_item_friends</item>
-        <item>@string/nav_item_notifications</item>
+        <item>@string/nav_item_filters</item>
+        <item>@string/nav_item_skin</item>
     </string-array>
 
-    <string name="title_messages">Messages</string>
-    <string name="title_friends">Friends</string>
+    <string name="title_change_skin">Change skin</string>
+    <string name="title_filters">Filters</string>
     <string name="title_home">Home</string>
 
     <!-- TODO: Remove or change this placeholder text -->
